+++ MNE/fiff/fiff_raw_data.h
@@ -129,7 +129,7 @@ public:
     *
     * @return true if succeeded, false otherwise
     */
-    bool read_raw_segment(MatrixXf*& data, MatrixXf*& times, fiff_int_t from = -1, fiff_int_t to = -1, MatrixXi sel = defaultMatrixXi);
+    bool read_raw_segment(MatrixXd*& data, MatrixXd*& times, fiff_int_t from = -1, fiff_int_t to = -1, MatrixXi sel = defaultMatrixXi);
 
 
     //=========================================================================================================
@@ -146,7 +146,7 @@ public:
     *
     * @return true if succeeded, false otherwise
     */
-    bool read_raw_segment_times(MatrixXf*& data, MatrixXf*& times, float from, float to, MatrixXi sel = defaultMatrixXi)
+    bool read_raw_segment_times(MatrixXd*& data, MatrixXd*& times, float from, float to, MatrixXi sel = defaultMatrixXi)
     {
         //
         //   Convert to samples
@@ -166,9 +166,9 @@ public:
     FiffInfo* info;
     fiff_int_t first_samp;
     fiff_int_t last_samp;
-    MatrixXf   cals;
+    MatrixXd   cals;
     QList<FiffRawDir> rawdir;
-    MatrixXf*  proj;
+    MatrixXd*  proj;
     FiffCtfComp comp;
 };
